brian and koji feedback
