+++ app/src/organisms/GripperWizardFlows/getGripperWizardSteps.ts
@@ -9,19 +9,11 @@ import {
   SUCCESSFULLY_DETACHED,
   SUCCESSFULLY_CALIBRATED,
 } from './constants'
-import { useFilterWizardStepsFrom } from '../../resources/wizards/hooks'
-
-import type { Subsystem } from '@opentrons/api-client'
 import type { GripperWizardStep, GripperWizardFlowType } from './types'
 
-export const useFilteredGripperWizardSteps = (
-  flowType: GripperWizardFlowType,
-  subsystem: Subsystem
-): GripperWizardStep[] =>
-  useFilterWizardStepsFrom(getGripperWizardSteps(flowType), subsystem)
-
 export const getGripperWizardSteps = (
-  flowType: GripperWizardFlowType
+  flowType: GripperWizardFlowType,
+  requiresFirmwareUpdate: boolean
 ): GripperWizardStep[] => {
   switch (flowType) {
     case GRIPPER_FLOW_TYPES.RECALIBRATE: {
@@ -40,7 +32,7 @@ export const getGripperWizardSteps = (
       ]
     }
     case GRIPPER_FLOW_TYPES.ATTACH: {
-      return [
+      const ALL_STEPS = [
         { section: SECTIONS.BEFORE_BEGINNING },
         { section: SECTIONS.MOUNT_GRIPPER },
         { section: SECTIONS.FIRMWARE_UPDATE },
@@ -56,6 +48,10 @@ export const getGripperWizardSteps = (
           successfulAction: SUCCESSFULLY_ATTACHED_AND_CALIBRATED,
         },
       ]
+
+      return requiresFirmwareUpdate
+        ? ALL_STEPS
+        : ALL_STEPS.filter(step => step.section !== SECTIONS.FIRMWARE_UPDATE)
     }
     case GRIPPER_FLOW_TYPES.DETACH: {
       return [
