+++ app/src/organisms/PipetteWizardFlows/__tests__/getPipetteWizardStepsForProtocol.test.tsx
@@ -4,7 +4,7 @@ import {
   mockAttachedPipetteInformation,
 } from '../../../redux/pipettes/__fixtures__'
 import { FLOWS, SECTIONS } from '../constants'
-import { getPipetteWizardStepsForProtocol } from '../useFilteredPipetteWizardSteps'
+import { getPipetteWizardStepsForProtocol } from '../getPipetteWizardStepsForProtocol'
 
 import type { PipetteWizardStep } from '../types'
 
@@ -36,14 +36,20 @@ describe('getPipetteWizardStepsForProtocol', () => {
             mount: 'left',
           },
         ],
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
   it('returns an empty array when there is no pipette attached and no pipette is needed', () => {
     const mockFlowSteps = [] as PipetteWizardStep[]
     expect(
-      getPipetteWizardStepsForProtocol({ left: null, right: null }, [], LEFT)
+      getPipetteWizardStepsForProtocol(
+        { left: null, right: null },
+        [],
+        LEFT,
+        true
+      )
     ).toStrictEqual(mockFlowSteps)
   })
   it('returns the calibration flow only when correct pipette is attached but there is no pip cal data', () => {
@@ -75,7 +81,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
           } as any,
         },
         [{ id: '123', pipetteName: 'p1000_single_flex', mount: 'right' }],
-        RIGHT
+        RIGHT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -123,7 +130,52 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         mockSingleMountPipetteAttached,
         mockPipettesInProtocolMulti as any,
-        LEFT
+        LEFT,
+        true
+      )
+    ).toStrictEqual(mockFlowSteps)
+  })
+  it('returns the correct array of info when the attached pipette does not require a firmware update', () => {
+    const mockFlowSteps = [
+      {
+        section: SECTIONS.BEFORE_BEGINNING,
+        mount: LEFT,
+        flowType: FLOWS.DETACH,
+      },
+      {
+        section: SECTIONS.DETACH_PIPETTE,
+        mount: LEFT,
+        flowType: FLOWS.DETACH,
+      },
+      { section: SECTIONS.RESULTS, mount: LEFT, flowType: FLOWS.DETACH },
+      {
+        section: SECTIONS.MOUNT_PIPETTE,
+        mount: LEFT,
+        flowType: FLOWS.ATTACH,
+      },
+      { section: SECTIONS.RESULTS, mount: LEFT, flowType: FLOWS.ATTACH },
+      {
+        section: SECTIONS.ATTACH_PROBE,
+        mount: LEFT,
+        flowType: FLOWS.ATTACH,
+      },
+      {
+        section: SECTIONS.DETACH_PROBE,
+        mount: LEFT,
+        flowType: FLOWS.ATTACH,
+      },
+      {
+        section: SECTIONS.RESULTS,
+        mount: LEFT,
+        flowType: FLOWS.CALIBRATE,
+      },
+    ] as PipetteWizardStep[]
+    expect(
+      getPipetteWizardStepsForProtocol(
+        mockSingleMountPipetteAttached,
+        mockPipettesInProtocolMulti as any,
+        LEFT,
+        false
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -181,7 +233,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         { left: mock96ChannelAttachedPipetteInformation, right: null },
         mockPipettesInProtocolNotEmpty as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -239,7 +292,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         { left: mockAttachedPipetteInformation, right: null },
         mockPipetteInfo as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -297,7 +351,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         { left: null, right: mockAttachedPipetteInformation },
         mockPipetteInfo as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -364,7 +419,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
           right: mockAttachedPipetteInformation,
         },
         mockPipetteInfo as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -406,7 +462,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         { left: null, right: null },
         mockPipettesInProtocolNotEmpty as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
@@ -458,7 +515,8 @@ describe('getPipetteWizardStepsForProtocol', () => {
       getPipetteWizardStepsForProtocol(
         { left: null, right: null },
         mockPipetteInfo as any,
-        LEFT
+        LEFT,
+        true
       )
     ).toStrictEqual(mockFlowSteps)
   })
